requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
openai==1.3.0
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.21.0
googletrans==4.0.0rc1
requests==2.31.0
python-dotenv==1.0.0
twilio==8.10.0
speechrecognition==3.10.0
pydub==0.25.1
pyaudio==0.2.11
requests==2.31.0
